pyproject.toml
[tool.poetry]
name = "telegram-ig-interaction"
version = "0.1.0"
description = ""
authors = ["barbamento <lorenzo.alvisi@studio.unibo.it>"]
readme = "README.md"
packages = [{include = "telegram_ig_interaction"}]

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
pandas = "^2.1.1"
python-telegram-bot = "^20.6"
requests = "^2.31.0"
black = "^23.10.1"
facebook-sdk = "^3.1.0"
instagrapi = "^2.0.0"
pillow = "^10.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
